+++ src/lib.rs
@@ -1,5 +1,3 @@
-mod utils;
-
 pub mod contrast;
 pub mod fastica;
 pub mod ffi;
